requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
alembic>=1.10.0

httpx>=0.24.0
python-dotenv>=0.21.0

pydantic>=2.1.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

PyJWT>=2.8.0
openai>=1.0.0

psycopg2-binary>=2.9.0
